[FLOW-6] refactor: erd-page refactor
