+++ src/features/erd-page/index.ts
@@ -1,0 +1,1 @@
+export * from './erd-project-page.hoos';
