+++ src/page/erd-project.page.tsx
@@ -1,7 +1,6 @@
 'use client';
 
 import styled from '@emotion/styled';
-import { useCallback, useEffect, useLayoutEffect, useState } from 'react';
 
 import {
   ErdDrawTools,
@@ -11,95 +10,32 @@ import {
 import { Relationship } from '@/components/Relationship';
 import { Table } from '@/components/table';
 import { useDrawToolsStore } from '@/features/draw-tools';
+import { useCanvasSize, usePageMove, useLastTable } from '@/features/erd-page';
 import { ERDTable } from '@/features/erd-project';
 import { useERDProjectStore } from '@/providers';
 
 export function ErdProjectPage() {
-  const projectWidth = useERDProjectStore((state) => state.width);
-  const projectHeight = useERDProjectStore((state) => state.height);
-
-  const updateCanvasSize = useERDProjectStore(
-    (state) => state.updateCanvasSize,
-  );
+  const { projectWidth, projectHeight } = useCanvasSize();
+  const { lastTable, setLastTable } = useLastTable();
+  usePageMove();
 
   const tables = useERDProjectStore((state) => state.tables);
-
   const createTable = useERDProjectStore((state) => state.createTable);
   const updateTable = useERDProjectStore((state) => state.updateTable);
-
   const createRelation = useERDProjectStore((state) => state.createRelation);
 
   const mapping = useDrawToolsStore((state) => state.mapping);
   const setMapping = useDrawToolsStore((state) => state.setMapping);
-
   const entity = useDrawToolsStore((state) => state.entity);
   const setEntity = useDrawToolsStore((state) => state.setEntity);
-
   const cursor = useDrawToolsStore((state) => state.cursor);
-
-  const [lastTable, setLastTable] = useState<ERDTable | undefined>(undefined);
-
-  const [pageMove, setPageMove] = useState<boolean>(false);
-
-  const [initPagePosition, setInitPagePosition] = useState<{
-    left: number;
-    top: number;
-  }>({ left: -1, top: -1 });
-  const [scrollPagePosition, setScrollPagePosition] = useState<{
-    left: number;
-    top: number;
-  }>({ left: -1, top: -1 });
-
-  const projectMaxDistance = 300;
-
-  const handleResize = useCallback(() => {
-    if (
-      window.innerWidth > projectWidth ||
-      window.innerHeight > projectHeight
-    ) {
-      updateCanvasSize({
-        width: Math.max(window.innerWidth, projectWidth),
-        height: Math.max(window.innerHeight, projectHeight),
-      });
-    }
-  }, [projectWidth, projectHeight, updateCanvasSize]);
-
-  const onPositionChange = (id: string, pos: { left: number; top: number }) => {
-    const table = tables.find((t) => t.id === id);
-
-    if (cursor === 'ARROW' && table) {
-      const updatedTable = { ...table, left: pos.left, top: pos.top };
-      updateTable(updatedTable);
-
-      const updateProjectWidth = Math.max(
-        projectWidth,
-        updatedTable.left + updatedTable.width + projectMaxDistance,
-      );
-      const updateProjectHeight = Math.max(
-        projectHeight,
-        updatedTable.top + updatedTable.height + projectMaxDistance,
-      );
-
-      if (
-        updateProjectWidth > projectWidth ||
-        updateProjectHeight > projectHeight
-      ) {
-        updateCanvasSize({
-          width: updateProjectWidth,
-          height: updateProjectHeight,
-        });
-      }
-    }
-  };
 
   const displayClicked = (
     event: React.MouseEvent<HTMLDivElement, MouseEvent>,
   ) => {
     const { pageX, pageY } = event;
-
     if (cursor === 'ARROW' && entity === 'TABLE') {
       const tableCount = tables.length;
-
       createTable({
         id: tableCount.toString(),
         title: 'table',
@@ -109,87 +45,17 @@ export function ErdProjectPage() {
         height: 30,
         columns: [],
       });
-
       setEntity('NONE');
-
-      const updateProjectWidth = Math.max(
-        projectWidth,
-        pageX + projectMaxDistance + 50,
-      );
-      const updateProjectHeight = Math.max(
-        projectHeight,
-        pageY + projectMaxDistance + 30,
-      );
-
-      if (
-        updateProjectWidth > projectWidth ||
-        updateProjectHeight > projectHeight
-      ) {
-        updateCanvasSize({
-          width: updateProjectWidth,
-          height: updateProjectHeight,
-        });
-      }
     }
   };
 
-  useEffect(() => {
-    const handleMouseDown = (e: { pageX: number; pageY: number }) => {
-      const { pageX, pageY } = e;
-
-      if (cursor === 'POINTER' && !pageMove) {
-        setPageMove(true);
-        setInitPagePosition({ left: pageX, top: pageY });
-
-        setScrollPagePosition({
-          left: window.scrollX,
-          top: window.scrollY,
-        });
-      }
-    };
-
-    const handleMouseMove = (e: { pageX: number; pageY: number }) => {
-      if (!pageMove) return;
-
-      const { pageX, pageY } = e;
-
-      const dx = pageX - initPagePosition.left;
-      const dy = pageY - initPagePosition.top;
-
-      window.scrollTo({
-        left: scrollPagePosition.left - dx,
-        top: scrollPagePosition.top - dy,
-        behavior: 'instant',
-      });
-    };
-
-    const handleMouseUp = () => {
-      if (pageMove) setPageMove(false);
-    };
-
-    if (cursor === 'POINTER') {
-      window.addEventListener('mousedown', handleMouseDown);
-      window.addEventListener('mousemove', handleMouseMove);
-      window.addEventListener('mouseup', handleMouseUp);
-    } else {
-      window.removeEventListener('mousedown', handleMouseDown);
-      window.removeEventListener('mousemove', handleMouseMove);
-      window.removeEventListener('mouseup', handleMouseUp);
+  const onPositionChange = (id: string, pos: { left: number; top: number }) => {
+    const table = tables.find((t) => t.id === id);
+    if (cursor === 'ARROW' && table) {
+      const updatedTable = { ...table, left: pos.left, top: pos.top };
+      updateTable(updatedTable);
     }
-
-    return () => {
-      window.removeEventListener('mousedown', handleMouseDown);
-      window.removeEventListener('mousemove', handleMouseMove);
-      window.removeEventListener('mouseup', handleMouseUp);
-    };
-  }, [
-    cursor,
-    initPagePosition.left,
-    initPagePosition.top,
-    pageMove,
-    scrollPagePosition.left,
-    scrollPagePosition.top,
-  ]);
+  };
 
   const TableClick = (table: ERDTable) => {
     if (
@@ -213,18 +79,6 @@ export function ErdProjectPage() {
       }
     }
   };
-
-  useLayoutEffect(() => {
-    handleResize();
-  }, []);
-
-  useEffect(() => {
-    window.addEventListener('resize', handleResize);
-
-    return () => {
-      window.removeEventListener('resize', handleResize);
-    };
-  }, [handleResize]);
 
   return (
     <styles.displayWrapper
@@ -256,7 +110,6 @@ const styles = {
     height: ${({ $pos }) => `${$pos.height}px`};
     background: #2f2f2f;
   `,
-
   container: styled.div`
     position: fixed;
     display: flex;
